Show slow bundle on pub/sub
First test in dev mode. Just run
  $meteor run

Then open up console and run the following method to insert some documents
in the TestCol collection:

Meteor.call('myBatchJobFillDatabase', 10000)

To run the update job run the following command in the console:

Meteor.call('myBatchJobUpdate', "test-name")

This updates the collection and measure time to do a findOne() on the same
collection. Look at code for more details.

Compare the time results between dev and production.

To build for production:
$ARCHITECTURE=os.linux.x86_64
$mkdir -p /tmp/meteor/builds && meteor build /tmp/meteor/builds --architecture ${ARCHITECTURE}

Then launch the project with node version 8.9.3